trading_system/requirements.txt
django
djangorestframework
yfinance
dotenv
psycopg2-binary